backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
websockets==12.0
# 文档解析
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
Pillow==10.4.0
# RAG / 向量检索
scikit-learn==1.5.2
numpy==1.26.4
